requirements.txt
BeautifulSoup4
Pillow
selenium
psycopg2-binary
python-dotenv
requests
spacy
https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.7.0/de_core_news_md-3.7.0-py3-none-any.whl
